moved new aks-related CLI code into new folder
