+++ maro/cli/k8s/aks/template.json
